// backend/src/modules/auth/dto/company.dto.ts
export interface CompanyDto {
  companyName: string;
  legalBusinessName?: string;
  industry: string;
  companySize: string;
  website?: string;
  businessType: string;
  currencyCode: string;
  taxNumber?: string;
  country: string;
  stateProvince: string;
  city: string;
  postalCode: string;
  timezone: string;
}